move copy operation to factory
